refactor: move location list query from database to AddCardForm component
