+++ .env.example
@@ -1,3 +1,5 @@
 NEXT_PUBLIC_API_BASE_URL=
 NEXT_PUBLIC_API_CRM_PATH=/crm
 NEXT_PUBLIC_API_STATS_PATH=/stats
+NEXT_PUBLIC_API_LOCATIONS_PATH=/locations
+NEXT_PUBLIC_API_ANIMAL_TYPES_PATH=/animal_types
