+++ src/components/crm/AddCardCrm/PopUp/CurrentLocation.tsx
@@ -2,39 +2,20 @@
 
 import { useState } from "react";
 import { AddLocation } from "./AddLocation";
-import { useQuery } from "@tanstack/react-query";
-import { fetch } from "@/src/utils/api";
-
-const API_CRM_PATH = process.env.NEXT_PUBLIC_API_CRM_PATH;
-const API_LOCATIONS_PATH = process.env.NEXT_PUBLIC_API_LOCATIONS_PATH;
+import { Location } from "../AddCardForm";
 
 interface PropsCurrentLocation {
+  locationsData: Location[];
   onChange: (value: string) => void;
   onClose: () => void;
 }
 
-export interface Location {
-  id: number;
-  name: string;
-}
-
 export const CurrentLocation: React.FC<PropsCurrentLocation> = ({
+  locationsData,
   onChange,
   onClose,
 }) => {
   const [isAddingLocation, setIsAddingLocation] = useState(false);
-
-  const {
-    data: locationsData,
-    isLoading,
-    isError,
-  } = useQuery<Location[]>({
-    queryKey: ["locationsData"],
-    queryFn: () => fetch(`${API_CRM_PATH}${API_LOCATIONS_PATH}`),
-  });
-
-  if (isLoading) return <p>Завантаження даних...</p>;
-  if (isError) return <p>Помилка завантаження даних</p>;
 
   const handleSelectLocation = (location: string) => {
     onChange(location);
